v1.3.0 Refactor staff commands and improve error handling
The code changes refactor out a ton of code, most notably staff commands to a new utility file to enhance modularity. The file includes functions that handle various staff interactions, such as creating and deleting teams, adding and removing players, etc. Also, the code enhances error handling by notifying the bot owner whenever an error occurs.
