+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "eh-overlord",
-  "version": "1.1.0",
+  "version": "1.3.0",
   "description": "",
   "main": "bot.js",
   "scripts": {
